+++ src/components/Feedback/Loading/Loading.tsx
@@ -1,4 +1,4 @@
-import { FC } from 'react';
+import React, { FC } from 'react';
 import ReactLoading from 'react-loading';
 
 import { LoadingProps } from './Loading.types';