+++ src/components/Feedback/Modal/Modal.types.ts
@@ -1,0 +1,6 @@
+export type ModalStylesProps = {
+  show: boolean;
+  onClose: () => void;
+};
+
+export type ModalProps = ModalStylesProps;