+++ src/components/Feedback/index.ts
@@ -1,2 +1,3 @@
 export { Loading } from './Loading';
 export { Toast } from './Toast';
+export { Modal } from './Modal';