+++ src/components/Form/IconButton/IconButton.tsx
@@ -1,15 +1,14 @@
-import { FC } from 'react';
+import React, { FC } from 'react';
 
-import { Icon } from 'components/DataDisplay/Icon';
-import {
-  Sizes,
-  Variants as VariantsIcon,
-} from 'components/DataDisplay/Icon/Icon.types';
 import {
   iconVariantRuleForButtons,
   iconVariantsForButtons,
-} from 'hook/components/useIcons';
-
+} from '../../../hook/components/useIcons';
+import { Icon } from '../../DataDisplay/Icon';
+import {
+  Sizes,
+  Variants as VariantsIcon,
+} from '../../DataDisplay/Icon/Icon.types';
 import * as Styled from './IconButton.styled';
 import { IconButtonProps } from './IconButton.types';