+++ src/components/Form/IconButton/IconButton.types.ts
@@ -1,7 +1,6 @@
 import { ButtonHTMLAttributes } from 'react';
 
-import { Icons } from 'types/Icons';
-
+import { Icons } from '../../../types/Icons';
 import { Variants } from '../Button/Button.types';
 
 type Sizes = 'md' | 'lg';